Pulling from GitHub the change made by Hafizzan on the Exercise function
